improve initializing logic, which is needed
At the moment, everything seems to be in order, literally. Now, the main
app is doing some elm/iced structure where we emit messages and then we
have to handle it somewhere. This seems to work great.

There is one error right now in WebGL2 and it is something about buffer
bindings not 16 byte aligned. Seems like an easy fix. It is all coming
together.