fix mipmap shader alignment thing
At this stage, there is 50% chance of running the project on the web.
The issue seems to come from <canvas>/window not being initialized
correctly. Need to mimic how wgpu does it on their end because theirs
works 100% of the times. Beside this, there is also something wrong with
texture binding. Maybe something with max array texture size? Not very
sure, but something is definitely wrong.