move egui_renderer to renderer module
It seems like anything renderer should go to the render module. Also, in
the future, when egui would become an integral part of kdr,
egui_renderer logic should be inside the render function as well. But
for the moment, I am not sure what is the right approach. The main egui
UI has to inherit values from the main App so that it knows what to
change. Too bad.