decoupling states from main App
Borrow checker doesn't like the previous design. The problem with the
previous design is the way that egui draw is called. It is intended that
egui has access to the app state and mutate it. However, in redraw
request event, render context is already borrowed. If egui is drawn
there, it also has to borrow the App. Now, with all the states
decoupled, render context is borrowing from render context, egui drawing
is borrowing from state, and the app is untouched. This makes some
stuffs fairly awkward like how mouse input needs window and window is
not part of state. Whatever.